% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converter.R
\name{asSmoof}
\alias{asSmoof}
\title{Converts coco problem to smoof function.}
\usage{
asSmoof(problem)
}
\arguments{
\item{problem}{[\code{\link{CocoProblem}}]\cr
Coco problem.}
}
\value{
[\code{smoof_function}]
}
\description{
See title
}
