% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{coco_observers}
\alias{coco_observers}
\title{COCO Observers}
\format{
An object of class \code{character} of length 3.
}
\usage{
coco_observers
}
\description{
List of available COCO observers.
}
\keyword{datasets}
