% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{coco_suites}
\alias{coco_suites}
\title{COCO Suites}
\format{
An object of class \code{character} of length 8.
}
\usage{
coco_suites
}
\description{
List of available COCO suites.
}
\keyword{datasets}
