% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoBenchmarkOptimizer.R
\name{cocoBenchmarkOptimizer}
\alias{cocoBenchmarkOptimizer}
\title{Run optimizer on complete coco suite.}
\usage{
cocoBenchmarkOptimizer(optimizer, suite, ...)
}
\arguments{
\item{optimizer}{[\code{function(fn, init, problem, ...)}]\cr
Runs an optimizer on a coco problem, verly likely a mini-wrapper for your optimizer of choice.
\code{fn} is the R function to optimize, with signature \code{function(x)}, where \code{x} is the numeric vector to evaluate at.
\code{problem} is a \code{link{CocoProblem}} that the optimizer can access for box constraints, initial solution and other info (or ignore).
The dotargs \code{...} are further arguments from above that should be passed on to the optimizer.}

\item{suite}{[\code{\link{CocoSuite}}]\cr
Coco problem suite.}

\item{\ldots}{[any]\cr
Passed down to \code{optimizer}.}
}
\value{
[\code{list}]. List of results for individual \code{\link{cocoRunOptimizer}} calls.
  List is named with coco problem ids.
}
\description{
The function is mainly called for its side effect of coco logging of results on disk.
It is expected that the caller opens the suite with \code{\link{cocoOpenSuite}} and closes the 
suite with \code{\link{cocoCloseSuite}}.
}
