% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoCloseSuite.R
\name{cocoCloseSuite}
\alias{cocoCloseSuite}
\title{Closes coco problem.}
\usage{
cocoCloseSuite(suite)
}
\arguments{
\item{suite}{[\code{\link{CocoSuite}}]\cr
Coco problem suite.}
}
\value{
[\code{invisible(NULL)}]
}
\description{
Must be called after all evaluations on problem are finished.
}
