% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoEvaluateFunction.R
\name{cocoEvaluateFunction}
\alias{cocoEvaluateFunction}
\title{Evaluate a coco function at point x.}
\usage{
cocoEvaluateFunction(problem, x)
}
\arguments{
\item{problem}{[\code{\link{CocoProblem}}]\cr
Coco problem.}

\item{x}{[\code{numeric}]\cr
Point to evaluate at, of length \code{problem$nr.of.dimensions}.}
}
\value{
[\code{numeric}]. Objective value, of length \code{problem$nr.of.objectives}.
}
\description{
See title
}
