% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoProblemInitObserver.R
\name{cocoInitObserver}
\alias{cocoInitObserver}
\title{Init coco observer.}
\usage{
cocoInitObserver(observer.name = "bbob", algorithm.name = "R_algo",
  algorithm.info = "", result.folder = paste0("R_on_", observer.name))
}
\arguments{
\item{observer.name}{[\code{character(1)}]\cr
Name of observer.
Default is \dQuote{bbob}.}

\item{algorithm.name}{[\code{character(1)}]\cr
Name of the algorithm.
Default is \dQuote{R_algo}.}

\item{algorithm.info}{[\code{character(1)}]\cr
Optional description of the algorithm.
Default is the empty character string.}

\item{result.folder}{[\code{character(1)}]\cr
Directory for the observer to write the output.
If the directory already exists the observer will automatically append \dQuote{-001} to the name.
Default is \dQuote{R_on_<observer.name>}.}
}
\value{
[\code{CocoObserver}].
}
\description{
See title
}
