% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CocoObserver.R
\name{CocoObserver}
\alias{CocoObserver}
\title{Coco observer object.}
\description{
The S3 class is a list which stores these elements:
\describe{
  \item{observer.name [\code{character(1)}]}{Name of the observer.}
  \item{algorithm.name [\code{character(1)}]}{Name of the algorithm interfaced by R.}
  \item{algorithm.info [\code{character(1)}]}{Textual description of \code{algorithm.name}.}
  \item{number.target.triggers [\code{integer(1)}]}{The number of targets between each 10**i and 10^(i+1)}
  \item{number.evaluation.triggers [\code{integer(1)}]}{The number of triggers between each 10**i and 10^(i+1) evaluation number.}
  \item{base.evaluation.triggers [\code{integer}]}{Defines the base evaluations used to produce an additional evaluation-based logging.}
  \item{precision.x [\code{integer(1)}]}{Output precision for decision variables.}
  \item{precision.f [\code{integer(1)}]}{Output precision for function values.}
  \item{result.folder [\code{character(1)}]}{Directory for the observer to write the output.}
  \item{observer.extptr}{External pointer to C object of coco observer.}
}
}
