% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoOpenSuite.R
\name{cocoOpenSuite}
\alias{cocoOpenSuite}
\title{Open a coco problem suite.}
\usage{
cocoOpenSuite(suite.name = "bbob", observer.name = "bbob")
}
\arguments{
\item{suite.name}{[\code{character(1)}]\cr
Name of coco suite.
Default is \dQuote{bbob}.}

\item{observer.name}{[\code{character(1)}]\cr
Name of observer.
Default is \dQuote{bbob}.}
}
\value{
[\code{\link{CocoSuite}}].
}
\description{
Must be called before problems can be accessed.
}
