% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco_problem.R
\name{CocoProblem}
\alias{CocoProblem}
\title{COCO Problem R6 Class}
\description{
Represents a COCO optimization problem.
This is simply a mathematical test function you can evaluate at a given points.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{problem_ptr}}{\code{(externalptr)}\cr
Pointer to the C problem object}

\item{\code{problem_idx}}{(\code{integer(1)})\cr
The index of the problem in the suite}

\item{\code{name}}{(\code{character(1)})\cr
Problem name}

\item{\code{id}}{(\code{integer(1)})\cr
Problem ID}

\item{\code{type}}{(\code{integer(1)})\cr
Problem type}

\item{\code{dim}}{(\code{integer(1)})\cr
Problem dimension}

\item{\code{fun_idx}}{(\code{integer(1)})\cr
Function index}

\item{\code{n_obj}}{(\code{integer(1)})\cr
Number of objectives}

\item{\code{n_constr}}{(\code{integer(1)})\cr
Number of constraints}

\item{\code{n_int}}{(\code{integer(1)})\cr
Number of integer variables}

\item{\code{lower}}{(\code{numeric(dim)})\cr
Lower bounds of decision space}

\item{\code{upper}}{(\code{numeric(dim)})\cr
Upper bounds of decision space}

\item{\code{target}}{(\code{numeric(1)})\cr
Target value for the first objective}

\item{\code{fupper}}{(\code{numeric(1)})\cr
Upper bounds of objective space, only for multi-objective problems}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CocoProblem-new}{\code{CocoProblem$new()}}
\item \href{#method-CocoProblem-eval}{\code{CocoProblem$eval()}}
\item \href{#method-CocoProblem-print}{\code{CocoProblem$print()}}
\item \href{#method-CocoProblem-clone}{\code{CocoProblem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new COCO Problem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblem$new(suite, problem_idx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{suite}}{(\link{CocoSuite})\cr
The COCO suite}

\item{\code{problem_idx}}{(\code{integer(1)})\cr
The index of the problem in the suite}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblem-eval"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblem-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluate the problem function at a given point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblem$eval(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{numeric(dim)})\cr
The point to evaluate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{numeric(n_obj)})\cr
The function value(s) at x
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblem-print"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblem-print}{}}}
\subsection{Method \code{print()}}{
Printer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblem$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
