% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoProblemFree.R
\name{cocoProblemFree}
\alias{cocoProblemFree}
\title{Free the given coco problem.}
\usage{
cocoProblemFree(problem)
}
\arguments{
\item{problem}{[\code{\link{CocoProblem}}]\cr
Coco problem.}
}
\value{
[\code{invisible(NULL)}]
}
\description{
This function is needed if one manually iterates over the problems of a
\code{\link{CocoSuite}} and wraps these with a \code{CocoObserver} since the
current observer needs to be closed before a new one is stared. This function
is used internally by \code{\link{cocoSuiteRunOptimizer}} and will rarely be
needed.
}
