% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoProblemGetBestObservedFValue.R
\name{cocoProblemGetBestObservedFValue}
\alias{cocoProblemGetBestObservedFValue}
\title{Get best observed value of the first function objective.}
\usage{
cocoProblemGetBestObservedFValue(problem)
}
\arguments{
\item{problem}{[\code{\link{CocoProblem}}]\cr
Coco problem.}
}
\value{
[\code{numeric(1)}]
}
\description{
See title
}
