% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoProblemGetEvaluations.R
\name{cocoProblemGetEvaluations}
\alias{cocoProblemGetEvaluations}
\title{Get number of function evaluations that were performed for a coco problem.}
\usage{
cocoProblemGetEvaluations(problem)
}
\arguments{
\item{problem}{[\code{\link{CocoProblem}}]\cr
Coco problem.}
}
\value{
[\code{integer(1)}]
}
\description{
See title
}
