% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoProblemGetInitialSolution.R
\name{cocoProblemGetInitialSolution}
\alias{cocoProblemGetInitialSolution}
\title{Get initial feasible solution for a coco problem.}
\usage{
cocoProblemGetInitialSolution(problem)
}
\arguments{
\item{problem}{[\code{\link{CocoProblem}}]\cr
Coco problem.}
}
\value{
[\code{numeric}]. Objective value, of length \code{problem$nr.of.objectives}.
}
\description{
See title
}
