% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoProblemIsFinalTargetHit.R
\name{cocoProblemIsFinalTargetHit}
\alias{cocoProblemIsFinalTargetHit}
\title{Check if target is already hit.}
\usage{
cocoProblemIsFinalTargetHit(problem)
}
\arguments{
\item{problem}{[\code{\link{CocoProblem}}]\cr
Coco problem.}
}
\value{
[\code{logical(1)}]
}
\description{
See title
}
