% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco_problem_noisy.R
\name{CocoProblemNoisy}
\alias{CocoProblemNoisy}
\title{COCO Problem R6 Class}
\description{
Represents a COCO optimization problem.
}
\section{Super class}{
\code{\link[rcoco:CocoProblem]{rcoco::CocoProblem}} -> \code{CocoProblemNoisy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CocoProblemNoisy-new}{\code{CocoProblemNoisy$new()}}
\item \href{#method-CocoProblemNoisy-eval}{\code{CocoProblemNoisy$eval()}}
\item \href{#method-CocoProblemNoisy-clone}{\code{CocoProblemNoisy$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rcoco" data-topic="CocoProblem" data-id="print"><a href='../../rcoco/html/CocoProblem.html#method-CocoProblem-print'><code>rcoco::CocoProblem$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblemNoisy-new"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblemNoisy-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new COCO Problem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblemNoisy$new(fun_idx, dim)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{suite}}{The COCO suite}

\item{\code{problem_idx}}{The index of the problem in the suite}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblemNoisy-eval"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblemNoisy-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluate the problem at a given point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblemNoisy$eval(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The point to evaluate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The function value at x
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblemNoisy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblemNoisy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblemNoisy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
