% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoSetLogLevel.R
\name{cocoSetLogLevel}
\alias{cocoSetLogLevel}
\title{Set log level for coco platform.}
\usage{
cocoSetLogLevel(level = "info")
}
\arguments{
\item{level}{[\code{character(1)}]\cr
Log level, can be \dQuote{debug}, \dQuote{info}, \dQuote{warning}, \dQuote{error}.
Default is \dQuote{info}.}
}
\value{
[\code{invisible(NULL)}]
}
\description{
See title
}
