% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco_suite.R
\name{CocoSuite}
\alias{CocoSuite}
\title{COCO Suite R6 Class}
\description{
R6 class for creating and managing COCO optimization suites.
These are collections of optimization problems.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{(\code{character(1)})\cr
The name of the suite}

\item{\code{instance}}{(\code{character(1)})\cr
The instance identifier}

\item{\code{suite_ptr}}{(\code{externalptr})\cr
Pointer to the C suite object}

\item{\code{observer_ptr}}{(\code{externalptr})\cr
Pointer to the C observer object}

\item{\code{n_problems}}{(\code{integer(1)})\cr
Number of problems in the suite}

\item{\code{data}}{(\code{data.frame()})\cr
Data frame with problem information}

\item{\code{observer_name}}{(\code{character(1)})\cr
The name of the COCO observer}

\item{\code{observer_options}}{(\code{character(1)})\cr
Options for the COCO observer}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CocoSuite-new}{\code{CocoSuite$new()}}
\item \href{#method-CocoSuite-print}{\code{CocoSuite$print()}}
\item \href{#method-CocoSuite-clone}{\code{CocoSuite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoSuite-new"></a>}}
\if{latex}{\out{\hypertarget{method-CocoSuite-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new COCO Suite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoSuite$new(
  name,
  instance = "year: 2009",
  observer_name = NULL,
  observer_options = ""
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{(\code{character(1)})\cr
The name of the suite.
Valid values are \link{coco_suites}.}

\item{\code{instance}}{(\code{character(1)})\cr
The instance identifier}

\item{\code{observer_name}}{(\code{character(1)})\cr
The name of the COCO observer
Valid values are \link{coco_observers}.}

\item{\code{observer_options}}{(\code{character(1)})\cr
Options for the COCO observer}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoSuite-print"></a>}}
\if{latex}{\out{\hypertarget{method-CocoSuite-print}{}}}
\subsection{Method \code{print()}}{
Printer
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoSuite$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoSuite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CocoSuite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoSuite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
