% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoSuiteGetNextProblem.R
\name{cocoSuiteGetNextProblem}
\alias{cocoSuiteGetNextProblem}
\title{Get next coco problem.}
\usage{
cocoSuiteGetNextProblem(suite)
}
\arguments{
\item{suite}{[\code{\link{CocoSuite}}]\cr
Coco problem suite.}
}
\value{
[\code{\link{CocoProblem}}].
}
\description{
Can be used to iterate through a complete coco problem suite.
}
