% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoSuiteGetNumberOfProblems.R
\name{cocoSuiteGetNumberOfProblems}
\alias{cocoSuiteGetNumberOfProblems}
\title{Get the number of problems in the given suite.}
\usage{
cocoSuiteGetNumberOfProblems(suite)
}
\arguments{
\item{suite}{[\code{\link{CocoSuite}}]\cr
Coco problem suite.}
}
\value{
[\code{integer(1)}]
}
\description{
See title
}
