% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/cocoSuiteGetProblemByFunctionDimensionInstance.R
\name{cocoSuiteGetProblemByFunDimInst}
\alias{cocoSuiteGetProblemByFunDimInst}
\title{Get specific coco problem.}
\usage{
cocoSuiteGetProblemByFunDimInst(suite, f, d, i)
}
\arguments{
\item{suite}{[\code{\link{CocoSuite}}]\cr
Coco problem suite.}

\item{f}{[\code{integer(1)}]\cr
Function number. Starts at 1.}

\item{d}{[\code{integer(1)}]\cr
Dimension.}

\item{i}{[\code{integer(1)}]\cr
Instance number. Starts at 1.}
}
\value{
[\code{\link{CocoProblem}}].
}
\description{
NB: If the problem index is too large, coco might just issue a warning and return a different
problem.
}
