% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco_suite_noisy.R
\name{CocoSuiteNoisy}
\alias{CocoSuiteNoisy}
\title{COCO Suite R6 Class}
\description{
R6 class for creating and managing COCO optimization suites.
}
\section{Super class}{
\code{\link[rcoco:CocoSuite]{rcoco::CocoSuite}} -> \code{CocoSuiteNoisy}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CocoSuiteNoisy-new}{\code{CocoSuiteNoisy$new()}}
\item \href{#method-CocoSuiteNoisy-clone}{\code{CocoSuiteNoisy$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="rcoco" data-topic="CocoSuite" data-id="print"><a href='../../rcoco/html/CocoSuite.html#method-CocoSuite-print'><code>rcoco::CocoSuite$print()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoSuiteNoisy-new"></a>}}
\if{latex}{\out{\hypertarget{method-CocoSuiteNoisy-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new COCO Suite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoSuiteNoisy$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoSuiteNoisy-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CocoSuiteNoisy-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoSuiteNoisy$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
