% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rcoco-package}
\alias{rcoco}
\alias{rcoco-package}
\title{rcoco: R Connector for the COCO Framework for Black-Box Optimization}
\description{
This Provides access to the COCO framework for black-box optimization. This includes multiple benchmark suites and functions for single and multi-objective optimization. For more information see the COCO website.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/berndbischl/rcoco/}
  \item \url{https://coco.gforge.inria.fr/}
  \item Report bugs at \url{https://github.com/berndbischl/rcoco/issues}
}

}
\author{
\strong{Maintainer}: Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})

}
