% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{coco_set_log_level}
\alias{coco_set_log_level}
\title{Set COCO Log Level}
\usage{
coco_set_log_level(log_level)
}
\arguments{
\item{log_level}{\verb{[character(1)]}\cr
A character string specifying the log level.
Valid values are "error", "warning", "info", "debug", or "" (empty string to not change the level).}
}
\value{
\verb{[character(1)]}\cr
The previous log level as a character string.
}
\description{
Sets the logging level for COCO operations.
}
