% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoOpenSuite.R
\name{cocoOpenSuite}
\alias{cocoOpenSuite}
\title{Open a coco problem suite.}
\usage{
cocoOpenSuite(name = "bbob", result.folder = "exdata/", year = 2017L,
  instances = NULL, dims = NULL, dim.inds = NULL, fun.inds = NULL,
  inst.inds = NULL)
}
\arguments{
\item{name}{[\code{character(1)}]\cr
Name of coco problem suite.
Default is \dQuote{bbob}.}

\item{result.folder}{[\code{character(1)}]\cr
Prefix for the folder the actual coco output is written to.}

\item{year}{[\code{integer(1)}]\cr
Year of the corresponding BBOB workshop, defining the instances used in that year's benchmark.
Default is 2017.}

\item{instances}{[\code{integer}]\cr
Instances for the suite, starting from 1.
Default is all instances chosen in BBOB in \code{year}.}

\item{dims}{[\code{integer}]\cr
Selected dimensions.
Default is all.}

\item{dim.inds}{[\code{integer}]\cr
Selected dimension indices to keep in the suite, starting from 1.
Default is all.}

\item{fun.inds}{[\code{integer}]\cr
Selected function indices to keep in the suite, starting from 1.
Default is all.}

\item{inst.inds}{[\code{integer}]\cr
Selected instance indices to keep in the suite, starting from 1.
Default is all.}
}
\value{
[\code{\link{CocoSuite}}].
}
\description{
Must be called before problems can be accessed.
For further info on suites, instances and options please see \url{http://numbbo.github.io/coco-doc/C/}.

The suite contains a collection of problems constructed by a Cartesian product of the suite's
optimization functions, dimensions and instances. The functions and dimensions are defined by the
suite name, while the instances are defined with the \code{instances} parameter.
The suite can be filtered by specifying functions, dimensions and instances.

If both year and instances appear in the \code{suite.instance} string, only the first one
is taken into account.
If no \code{suite.instance} is given, it defaults to the year of the current BBOB workshop.


If both dimensions and dimension_indices appear in the \code{suite.options} string,
only the first one is taken into account.
}
