% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco_problem.R
\name{CocoProblem}
\alias{CocoProblem}
\title{COCO Problem R6 Class}
\description{
R6 class for creating and managing COCO optimization problems.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{problem_ptr}}{Internal pointer to the C problem object}

\item{\code{problem_idx}}{The index of the problem in the suite}

\item{\code{name}}{Problem name}

\item{\code{id}}{Problem ID}

\item{\code{type}}{Problem type}

\item{\code{dim}}{Problem dimension}

\item{\code{n_obj}}{Number of objectives}

\item{\code{n_constr}}{Number of constraints}

\item{\code{n_int}}{Number of integer variables}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CocoProblem-new}{\code{CocoProblem$new()}}
\item \href{#method-CocoProblem-eval}{\code{CocoProblem$eval()}}
\item \href{#method-CocoProblem-print}{\code{CocoProblem$print()}}
\item \href{#method-CocoProblem-clone}{\code{CocoProblem$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblem-new"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblem-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new COCO Problem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblem$new(suite, problem_idx)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{suite}}{The COCO suite}

\item{\code{problem_idx}}{The index of the problem in the suite}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblem-eval"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblem-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluate the problem at a given point
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblem$eval(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The point to evaluate}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The function value at x
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblem-print"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblem-print}{}}}
\subsection{Method \code{print()}}{
Print method for the COCO Problem
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblem$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoProblem-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CocoProblem-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoProblem$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
