% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CocoProblem.R
\name{CocoProblem}
\alias{CocoProblem}
\title{Coco problem object.}
\description{
The S3 class is a list which stores these elements:
\describe{
  \item{id [\code{character(1)}]}{Short id of problem.}
  \item{index [\code{integer(1)}]}{Index of problem in suite.}
  \item{name [\code{character(1)}]}{Longer name of problem.}
  \item{nr.of.objectives [\code{integer(1)}}{Number of objectives.}
  \item{dimension [\code{integer(1)}}{Dimension of decision space.}
  \item{nr.of.constraints [\code{integer(1)}}{Number of constraints.}
  \item{lower [\code{numeric}]}{Lower box constraints.}
  \item{upper [\code{numeric}]}{upper box constraints.}
  \item{extptr}{External pointer to C object of coco problem.}
}
}
