% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoProblemAddObserver.R
\name{cocoProblemAddObserver}
\alias{cocoProblemAddObserver}
\title{Wrap coco problem with coco observer.}
\usage{
cocoProblemAddObserver(problem, observer)
}
\arguments{
\item{problem}{[\code{\link{CocoProblem}}]\cr
Coco problem.}

\item{observer}{[\code{CocoObserver}]\cr
Coco observer generated via \code{\link{cocoInitObserver}}.}
}
\value{
[\code{\link{CocoProblem}}].
}
\description{
See title
}
