% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoProblemFree.R
\name{cocoProblemFree}
\alias{cocoProblemFree}
\title{Free the given coco problem}
\usage{
cocoProblemFree(problem)
}
\arguments{
\item{problem}{[\code{\link{CocoProblem}}]\cr
Coco problem.}
}
\value{
[\code{invisible(NULL)}]
}
\description{
See title
}
