% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CocoSuite.R
\name{CocoSuite}
\alias{CocoSuite}
\title{Coco problem suite object.}
\description{
The S3 class is a list which stores these elements:
\describe{
  \item{suite.name [\code{character(1)}]}{Name of suite.}
  \item{suite.extptr}{External pointer to C object of coco suite.}
  \item{observer.name [\code{character(1)}]}{Name of observer.}
  \item{observer.extptr}{External pointer to C object of coco observer.}
}
}
