% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coco_suite.R
\name{CocoSuite}
\alias{CocoSuite}
\title{COCO Suite R6 Class}
\description{
R6 class for creating and managing COCO optimization suites.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{The name of the suite}

\item{\code{instance}}{The instance identifier}

\item{\code{suite_ptr}}{Internal pointer to the C suite object}

\item{\code{n_problems}}{Number of problems in the suite}

\item{\code{data}}{Data frame with problem information}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CocoSuite-new}{\code{CocoSuite$new()}}
\item \href{#method-CocoSuite-print}{\code{CocoSuite$print()}}
\item \href{#method-CocoSuite-clone}{\code{CocoSuite$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoSuite-new"></a>}}
\if{latex}{\out{\hypertarget{method-CocoSuite-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new COCO Suite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoSuite$new(name, instance)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{The name of the suite}

\item{\code{instance}}{The instance identifier}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoSuite-print"></a>}}
\if{latex}{\out{\hypertarget{method-CocoSuite-print}{}}}
\subsection{Method \code{print()}}{
Print method for the COCO Suite
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoSuite$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CocoSuite-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CocoSuite-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CocoSuite$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
