% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoSuiteGetAllProblems.R
\name{cocoSuiteGetAllProblems}
\alias{cocoSuiteGetAllProblems}
\title{Get all problem of a coco suite.}
\usage{
cocoSuiteGetAllProblems(suite)
}
\arguments{
\item{suite}{[\code{\link{CocoSuite}}]\cr
Coco problem suite.}
}
\value{
[\code{list} of \code{\link{CocoProblem}}]. Named with coco problem ids.
}
\description{
Iterates through all problems by calling \code{\link{cocoSuiteGetNextProblem}}.
}
