% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoSuiteGetNextProblem.R
\name{cocoSuiteGetNextProblem}
\alias{cocoSuiteGetNextProblem}
\title{Get next coco problem.}
\usage{
cocoSuiteGetNextProblem(suite, observer)
}
\arguments{
\item{suite}{[\code{\link{CocoSuite}}]\cr
Coco problem suite.}

\item{observer}{[\code{character(1)}]\cr
Coco observer.}
}
\value{
problem [\code{\link{CocoProblem}} | \code{NULL}]\cr
  Coco problem. \code{NULL} is returned if no problems are left.
}
\description{
Can be used to iterate through a complete coco problem suite.
}
