% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cocoSuiteGetProblem.R
\name{cocoSuiteGetProblem}
\alias{cocoSuiteGetProblem}
\title{Get specific coco problem.}
\usage{
cocoSuiteGetProblem(suite, index)
}
\arguments{
\item{suite}{[\code{\link{CocoSuite}}]\cr
Coco problem suite.}

\item{index}{[\code{integer(1)}]\cr
Problem index.}
}
\value{
[\code{\link{CocoProblem}}].
}
\description{
NB: If the problem index is too large, coco might just issue a warning and return a different
problem.
}
