% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{rcoco-package}
\alias{rcoco}
\alias{rcoco-package}
\title{rcoco: Black-Box Optimization Toolkit}
\description{
Features highly configurable search spaces via the 'paradox' package and optimizes every user-defined objective function. The package includes several optimization algorithms e.g. Random Search, Iterated Racing, Bayesian Optimization (in 'mlr3mbo') and Hyperband (in 'mlr3hyperband'). bbotk is the base package of 'mlr3tuning', 'mlr3fselect' and 'miesmuschel'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bbotk.mlr-org.com}
  \item \url{https://github.com/mlr-org/bbotk}
  \item Report bugs at \url{https://github.com/mlr-org/bbotk/issues}
}

}
\author{
\strong{Maintainer}: Bernd Bischl \email{bernd_bischl@gmx.net} (\href{https://orcid.org/0000-0001-6002-6980}{ORCID})

}
